requirements.txt
langchain==1.2.0
langchain_openai==1.1.6
faiss-cpu==1.8.0
matplotlib==3.10.8
PyPDF2==3.0.1
python-dotenv==1.2.1
streamlit==1.50.0
